refactor: extract TeamRecordRepository (streak, season-start, team game/season records); add memoization test
